requirements.txt
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
future==1.0.0
geocoder==1.38.1
idna==3.6
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
multiprocess==0.70.16
nest-asyncio==1.6.0
np==1.0.2
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pathos==0.3.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pyzmq==25.1.2
ratelim==0.1.6
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
urllib3==2.2.1
wcwidth==0.2.13
